backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic-settings==2.7.*
pytigergraph==1.8.*
httpx==0.28.*
playwright==1.50.*
groq==0.18.*
python-dotenv==1.1.*
